Relocate deprecated methods to its own file
Wow. At last, a practical use for traits!
