+++ Idno/Core/Idno.php
@@ -11,10 +11,12 @@ namespace Idno\Core {
 
     use Idno\Common\Page;
     use Idno\Entities\User;
-
+    
     class Idno extends \Idno\Common\Component
     {
-
+        // Install Idno deprecated functions
+        use Deprecated\Idno;
+    
         public $db;
         public $filesystem;
         public $config;
@@ -476,90 +478,6 @@ namespace Idno\Core {
             if (is_callable($listener)) {
                 $this->dispatcher->addListener($event, $listener, $priority);
             }
-        }
-
-        /**
-         * Registers a page handler for a given pattern, using Toro
-         * page handling syntax
-         *
-         * @deprecated
-         * @param string $pattern The pattern to match
-         * @param string $handler The name of the Page class that will serve this route
-         * @param bool $public If set to true, this page is always public, even on non-public sites
-         */
-        function addPageHandler($pattern, $handler, $public = false)
-        {
-            \Idno\Core\Idno::site()->logging()->warning("DEPRECATION WARNING: \Idno\Core\Idno::site()->routes()->addRoute()");
-            
-            return $this->routes()->addRoute($pattern, $handler, $public);
-        }
-
-        /**
-         * Registers a page handler for a given pattern, using Toro
-         * page handling syntax - and ensures it will be handled first
-         *
-         * @deprecated
-         * @param string $pattern The pattern to match
-         * @param string $handler The name of the Page class that will serve this route
-         * @param bool $public If set to true, this page is always public, even on non-public sites
-         */
-        function hijackPageHandler($pattern, $handler, $public = false)
-        {
-            \Idno\Core\Idno::site()->logging()->warning("DEPRECATION WARNING: \Idno\Core\Idno::site()->routes()->hijackRoute()");
-            
-            return $this->routes()->hijackRoute($pattern, $handler, $public);
-        }
-
-        /**
-         * Mark a page handler class as offering public content even on walled garden sites
-         * @deprecated
-         * @param $class
-         */
-        function addPublicPageHandler($class)
-        {
-            \Idno\Core\Idno::site()->logging()->warning("DEPRECATION WARNING: \Idno\Core\Idno::site()->routes()->addPublicRoute()");
-            
-            return $this->routes()->addPublicRoute($class);
-        }
-
-        /**
-         * Retrieve an array of walled garden page handlers
-         * @deprecated
-         * @return array
-         */
-        function getPublicPageHandlers()
-        {
-            \Idno\Core\Idno::site()->logging()->warning("DEPRECATION WARNING: \Idno\Core\Idno::site()->routes()->getPublicRoute()");
-            
-            return $this->routes()->getPublicRoute();
-        }
-
-        /**
-         * Does the specified page handler class represent a public page, even on walled gardens?
-         * @deprecated
-         * @param $class
-         * @return bool
-         */
-        function isPageHandlerPublic($class)
-        {
-            \Idno\Core\Idno::site()->logging()->warning("DEPRECATION WARNING: \Idno\Core\Idno::site()->routes()->isRoutePublic()");
-            
-            return $this->routes()->isRoutePublic($class);
-        }
-
-        /**
-         * Retrieves an instantiated version of the page handler class responsible for
-         * a particular page (if any). May also be a whole URL.
-         *
-         * @deprecated
-         * @param string $path_info The path, including the initial /, or the URL
-         * @return bool|\Idno\Common\Page
-         */
-        function getPageHandler($path_info)
-        {
-            \Idno\Core\Idno::site()->logging()->warning("DEPRECATION WARNING: \Idno\Core\Idno::site()->routes()->getRoute()");
-            
-            return $this->routes()->getRoute($path_info);
         }
 
         /**
